requirements.txt
# Web API
fastapi
uvicorn[standard]

# Database
pymysql

# Data ingestion (eski pipeline icin, Faz 2'de degisecek)
requests
pandas

# Environment
python-dotenv
